refactor: utils 디렉토리에 함수 모듈화
