+++ src/js/index.js
@@ -1,58 +1,12 @@
-const $ = selector => document.querySelector(selector);
+import $ from './utils/common/selector.js';
+import { store } from './utils/common/store.js';
+import updateMenuCount from './utils/CRUD/updateMenuCount.js';
+import addMenuName from './utils/CRUD/addMenuName.js';
+import updateMenuName from './utils/CRUD/updateMenuName.js';
+import removeMenuName from './utils/CRUD/removeMenuName.js';
 
 const App = () => {
-  const updateMenuCount = () => {
-    const menuCount = $('#espresso-menu-list').querySelectorAll('li').length;
-    $('.menu-count').innerText = `총 ${menuCount}개`;
-  };
-
-  const addMenuName = () => {
-    if (!$('#espresso-menu-name').value) {
-      alert('값을 입력해주세요.');
-      return;
-    }
-
-    const espressoMenuName = $('#espresso-menu-name').value;
-    const menuItemTemplate = espressoMenuName => {
-      return `
-        <li class="menu-list-item d-flex items-center py-2">
-          <span class="w-100 pl-2 menu-name">${espressoMenuName}</span>
-          <button
-            type="button"
-            class="bg-gray-50 text-gray-500 text-sm mr-1 menu-edit-button"
-          >
-            수정
-          </button>
-          <button
-            type="button"
-            class="bg-gray-50 text-gray-500 text-sm mr-1 menu-remove-button"
-          >
-            삭제
-          </button>
-        </li>`;
-    };
-    $('#espresso-menu-list').insertAdjacentHTML(
-      'beforeend',
-      menuItemTemplate(espressoMenuName),
-    );
-
-    updateMenuCount();
-    $('#espresso-menu-name').value = '';
-  };
-
-  const updateMenuName = e => {
-    const $menuName = e.target.closest('li').querySelector('.menu-name');
-    const updatedMenuName = prompt('메뉴명을 수정하세요', $menuName.innerText);
-    $menuName.innerText = updatedMenuName;
-  };
-
-  const removeMenuName = e => {
-    if (confirm('정말 삭제하시겠습니까?')) {
-      const menu = e.target.closest('li');
-      menu.remove();
-      updateMenuCount();
-    }
-  };
+  // 상태(=변할 수 있는 데이터) - 갯수, 메뉴명
 
   $('#espresso-menu-list').addEventListener('click', e => {
     if (e.target.classList.contains('menu-edit-button')) {
