+++ src/js/utils/common/store.js
@@ -1,0 +1,9 @@
+export const store = {
+  setData(menu) {
+    localStorage.setItem('menu', JSON.stringify(menu));
+  },
+
+  getData() {
+    localStorage.getItem('menu');
+  },
+};
